+++ src/Text/Jira/Parser/Shared.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-|
 Module      : Text.Jira.Parser.Shared
-Copyright   : © 2019–2020 Albert Krewinkel
+Copyright   : © 2019–2021 Albert Krewinkel
 License     : MIT
 
 Maintainer  : Albert Krewinkel <tarleb@zeitkraut.de>
@@ -12,6 +12,7 @@ Parsers whch are shared between multiple modules.
 -}
 module Text.Jira.Parser.Shared
   ( icon
+  , colorName
   ) where
 
 import Data.Char (isLetter)
@@ -60,3 +61,8 @@ otherIcon = try $ do
     "flag"    -> pure IconFlag
     "flagoff" -> pure IconFlagOff
     _         -> fail ("not a known emoji" ++ name)
+
+colorName :: Parsec Text u String
+colorName = many1 letter <|> hexColor
+  where
+    hexColor = (:) <$> option '#' (char '#') <*> count 6 hexDigit